Update dependencies, holding openai and ipython back
This constrains openai<1 for now. Showing how to work with and wrap
the old Python API is a central part of this project, so it's
unclear what will be done with that moving forward. The options are
to stay with <1 indefinitely, move up to 1, or support both
conditionally (which should usually be avoided, but for this
project it might be nice to show how to do so when needed). Note
that embeddings_utils is not present in openai>=1.

Less importantly, this also pins ipython back to ~8.12.3, which is
actually a downgrade. This is to support Python 3.8. Before, we
were installing an older version just on older Pythons, as we do
with numpy and scipy. For ipython, the reason to give newer Pythons
the newer version was that the convenient and popular way to
autocomplete a full line, by pressing right-arrow, was broken, and
the fix initially had not been backported. This is now fixed, due
to the backporting of https://github.com/ipython/ipython/pull/14080
to 8.12, which is in the 8.12.3 patch.