Regard HTTP 403 links as "alive", for now (#266)
This reconfigures markdown-link-check to treat both HTTP 200 and
HTTP 403 status codes as meaning the link is not broken. This has a
high risk of false detection. It is a temporary workaround for the
problem of pages on the OpenAI website (linked to from the readme)
returning HTTP 403 when accessed from GitHub Actions runners.